Updates Dockerfile names and base images to 1803
microsoft/windowsservercore:1803 has been released.

This patch replaces microsoft/windowsservercore:1709 as the base
image and renames the Dockerfile in a more uniform way.

The docker files can accept a BASE_IMAGE as an argument (by default, it
will use microsoft/windowsservercore:1803), which will be used when
building the images.

Fixes busybox and curl images (1803 comes with curl, which means the
curl image cannot be built without these changes).