Fixes netexec image
The netexec image is based on busybox, which sets the WORKDIR to C:\busybox. Because of this, the ADD command is currently adding the netexec.exe file into the busybox folder, and the entrypoint expects it to be in the root folder.

This patch fixes this issue by adding the file to the root folder.