Fixes guestbook application
The guestbook app deploys 3 types of containers: gb-frontend,
gb-redisslave, and redis. They all had various issues, causing the app
not to work properly:

1. DNS resolution: gb-frontend and gb-redisslave tries to connect to redis
and gb-redisslave through DNS name. But if DNS suffix is wrong, they
won't be able to connect. In the case of Kubernetes, host entries will
be added for redis-slave and and redis-master for the containers, so
they will be able to resolve the names.

2. run.ps1 was not being executed.

3. Redis was not being started using the added config value, starting in
protected mode, refusing all connections.