Updates go binary building and the webhook Docker image
The update is required for some Kubernetes E2E tests, in particular
by those requiring the /pods/attach handler.

This commit also adds the possibility to build executables from other
projects using their file "links" (files with the .golnk extension).
This will reduce the necessity to update the in-tree go code as often.