Cleans up golang source files
Previously, we used to build the golang source files from within this
repo, but since then we've introduced a mechanism in BuildImages.ps1
which allows it to follow links in order to build the binaries.

This patch removes most of the remaining golang source files, which
have no additional changes and replaces them with links to their
original files. One benefit to this would be that we won't have to
make sure that the local source files are in sync with the original
ones.

Additionally, this patch moves the hostname.go file to go_files/hostname
and adding a link instead, reducing a needless duplicate.

Removes orphan echoserver.go file, which was no longer used.