Add separate templates for staging job and main job.
The template for the main job ( kubernetes_template.json ) will
have pinned versions for Windows Image, CNI version and Docker version.
The reason behind the change is to protect the main job against
changes in these components that may affect the success of tests.

The staging job ( with template kubernetes_staging_template.json ) will
be used to experiment with new versions for CNI etc.