tests: Recreates go executables if Recreate is true
At the moment, the go executables are built only once; if the executable
is already found, they will not be rebuilt.

If the go sources are modified, the executables would need to be rebuilt
manually or deleted before recreating the images.

This commit passes the $Recreate flag to the BuildGoFiles function,
which will now rebuild the executables if true.