Removes zookeeper DockerImage
The image was used by a test that was deleted some time ago. (should create and stop Zookeeper, Nimbus and Storm worker servers, which existed in test/e2e/examples.go).

Since that test was deleted, the image is no longer needed.  The image was WIP, and its Dockerfile wasn't committed.