Updates agnhost, nettest, resource-consumer, webhook
The agnhost image has been approved in the k/k repo, and it includes
golang sources. We can build the binary by using a .golnk file.

The network connectivity bugfix also merged for nettest, so there is no
longer a reason to keep the source code in-tree.

The resource-consumer/controller image has been renamed to
resource-consumer-controller, and the sources have been moved, so the
.golnk needs to be updated as well.

The versions for the following images has been bumped:
resource-consumer, webhook.