Create test-known-issues.md
Create a document to centralise any known issues in the tests and the proposed fixes, as well as tracking exclusions and the reasoning behind them. 

We should centralise all known information about tests on Windows to have as a reference for any reviewer/user who might be interested in the failures. Similar to what we are doing with aks-engine versions used in testing.