Update prepulled Windows test containers
Forks prepull.yaml into version-specific manifests for each K8s version
we test on Windows nodes. The original prepull.yaml is left in place for
now and will be removed after k/k is updated to choose the
version-specific file.

Removes the initContainers specification, as this causes the prepull to
halt if some container can't be pulled.