Revert "run cmd/kubeadm/test: add env KUBEADM_PATH and build kubeadm.exe"
The commit that is being reverted was meant to build the kubeadm binary, so
it can be used to pass a few kubeadm-related unit tests. However, the modified
script is for E2E tests, not unit tests. The commit is not doing what was
intended, and it needlessly adds to the E2E job run times.

This reverts commit ac69bf750af1377f6ce31d79423f93bff9c5943c.