Parameterize k8s_unit_windows.ps1 to accept a list of packages
We want to be able to run the unit tests only on a subset of packages, reducing
the amount of total time spent by the job. For this, we need to be able to
specify a list of packages to test.

Added testPackages argument to the k8s_unit_windows.ps1 script.

ci-k8s-unit-test.sh can be configured with the TEST_PACKAGES env variable, which
will then pass to the k8s_unit_windows.ps1 script if given.